+++ apps/Standalone/src/templates/app/TemplatesStandaloneDesigner.tsx
@@ -1,35 +1,43 @@
 import { useMemo, type ReactNode } from 'react';
-import { TemplatesDataProvider } from '@microsoft/logic-apps-designer';
+import { TemplateFilters, TemplatesDataProvider } from '@microsoft/logic-apps-designer';
 import { environment, loadToken } from '../../environments/environment';
 import { DevToolbox } from '../components/DevToolbox';
 import type { RootState } from '../state/Store';
 import { TemplatesDesigner, TemplatesDesignerProvider } from '@microsoft/logic-apps-designer';
 import { useQuery } from '@tanstack/react-query';
 import { useSelector } from 'react-redux';
-import { BaseGatewayService, BaseTenantService, StandardConnectionService } from '@microsoft/logic-apps-shared';
 import {
+  BaseGatewayService,
+  StandardTemplateService,
+  BaseTenantService,
+  StandardConnectionService,
+  clone,
+  escapeSpecialChars,
+  isArmResourceId,
+  optional,
+  StandardOperationManifestService,
+} from '@microsoft/logic-apps-shared';
+import {
+  getConnectionStandard,
   useAppSettings,
   useConnectionsData,
   useCurrentObjectId,
   useCurrentTenantId,
   useWorkflowApp,
 } from '../../designer/app/AzureLogicAppsDesigner/Services/WorkflowAndArtifacts';
-import type { ConnectionsData } from '../../designer/app/AzureLogicAppsDesigner/Models/Workflow';
+import type { ConnectionAndAppSetting, ConnectionsData } from '../../designer/app/AzureLogicAppsDesigner/Models/Workflow';
 import type { WorkflowApp } from '../../designer/app/AzureLogicAppsDesigner/Models/WorkflowApp';
 import { ArmParser } from '../../designer/app/AzureLogicAppsDesigner/Utilities/ArmParser';
 import { StandaloneOAuthService } from '../../designer/app/AzureLogicAppsDesigner/Services/OAuthService';
-import { WorkflowUtility } from '../../designer/app/AzureLogicAppsDesigner/Utilities/Workflow';
+import { WorkflowUtility, addConnectionInJson, addOrUpdateAppSettings } from '../../designer/app/AzureLogicAppsDesigner/Utilities/Workflow';
 import { HttpClient } from '../../designer/app/AzureLogicAppsDesigner/Services/HttpClient';
-// import { useNavigate } from 'react-router-dom';
-// import type { Template, LogicAppsV2 } from '@microsoft/logic-apps-shared';
-// import { saveWorkflowStandard } from '../../designer/app/AzureLogicAppsDesigner/Services/WorkflowAndArtifacts';
-// import type { ParametersData } from '../../designer/app/AzureLogicAppsDesigner/Models/Workflow';
-import { useNavigate } from 'react-router-dom';
 import type { Template, LogicAppsV2, IWorkflowService } from '@microsoft/logic-apps-shared';
 import { saveWorkflowStandard } from '../../designer/app/AzureLogicAppsDesigner/Services/WorkflowAndArtifacts';
 import type { ParametersData } from '../../designer/app/AzureLogicAppsDesigner/Models/Workflow';
 import axios from 'axios';
+import type { ConnectionMapping } from '../../../../../libs/designer/src/lib/core/state/templates/workflowSlice';
 
+const workflowIdentifier = '#workflowname#';
 const LoadWhenArmTokenIsLoaded = ({ children }: { children: ReactNode }) => {
   const { isLoading } = useQuery(['armToken'], loadToken);
   return isLoading ? null : <>{children}</>;
@@ -37,48 +45,30 @@ const LoadWhenArmTokenIsLoaded = ({ children }: { children: ReactNode }) => {
 export const TemplatesStandaloneDesigner = () => {
   const theme = useSelector((state: RootState) => state.workflowLoader.theme);
   const { appId, isConsumption, workflowName: existingWorkflowName } = useSelector((state: RootState) => state.workflowLoader);
-  const { data: workflowAppData } = useWorkflowApp(appId as string);
+  const { data: workflowAppData } = useWorkflowApp(appId as string, isConsumption ? 'consumption' : 'standard');
   const canonicalLocation = WorkflowUtility.convertToCanonicalFormat(workflowAppData?.location ?? '');
   const { data: tenantId } = useCurrentTenantId();
   const { data: objectId } = useCurrentObjectId();
-  const { data: connectionsData } = useConnectionsData(appId);
+  const { data: originalConnectionsData } = useConnectionsData(appId);
   const { data: settingsData } = useAppSettings(appId as string);
-  const navigate = useNavigate();
 
-  // const navigate = useNavigate();
+  const connectionsData = useMemo(() => {
+    return JSON.parse(JSON.stringify(clone(originalConnectionsData ?? {})));
+  }, [originalConnectionsData]);
+
   const connectionReferences = WorkflowUtility.convertConnectionsDataToReferences(connectionsData);
-  const sanitizeParameterName = (parameterName: string, workflowName: string) =>
-    parameterName.replace('_#workflowname#', `_${workflowName}`);
 
   const createWorkflowCall = async (
     workflowName: string,
     workflowKind: string,
     workflowDefinition: LogicAppsV2.WorkflowDefinition,
-    _connectionsData: any,
+    connectionsMapping: ConnectionMapping,
     parametersData: Record<string, Template.ParameterDefinition>
   ) => {
     const workflowNameToUse = existingWorkflowName ?? workflowName;
-    const callBack = () => {
-      console.log('Created workflow, TODO: now redirect');
-      navigate('/');
-    };
     if (appId) {
       if (isConsumption) {
         console.log('Consumption is not ready yet!');
-        // await saveWorkflowConsumption({
-        //   id: appId,
-        //   name: workflowNameToUse,
-        //   type: "json", //TODO: figure out what this type is and replace it
-        //   kind: workflowKind,
-        //   properties: {
-        //     files: {
-        //       [Artifact.WorkflowFile]: workflow,
-        //       [Artifact.ParametersFile]: parametersData as ParametersData,
-        //       [Artifact.ConnectionsFile]: _connectionsData
-        //     },
-        //     health: {},
-        //   }
-        // }, workflow);
       } else {
         let sanitizedWorkflowDefinitionString = JSON.stringify(workflowDefinition);
         const sanitizedParameterData: ParametersData = {};
@@ -86,7 +76,7 @@ export const TemplatesStandaloneDesigner = () => {
         // Sanitizing parameter name & body
         Object.keys(parametersData).forEach((key) => {
           const parameter = parametersData[key];
-          const sanitizedParameterName = sanitizeParameterName(parameter.name, workflowName);
+          const sanitizedParameterName = replaceWithWorkflowName(parameter.name, workflowName);
           sanitizedParameterData[sanitizedParameterName] = {
             type: parameter.type,
             description: parameter?.description,
@@ -98,10 +88,21 @@ export const TemplatesStandaloneDesigner = () => {
           );
         });
 
+        const {
+          connectionsData: updatedConnectionsData,
+          settingProperties: updatedSettingProperties,
+          workflowJsonString: updatedWorkflowJsonString,
+        } = await updateConnectionsDataWithNewConnections(
+          connectionsData,
+          settingsData?.properties,
+          connectionsMapping,
+          sanitizedWorkflowDefinitionString,
+          workflowName
+        );
+        sanitizedWorkflowDefinitionString = updatedWorkflowJsonString;
+
         const workflow = {
           definition: JSON.parse(sanitizedWorkflowDefinitionString),
-          connectionReferences: undefined, //TODO: change this after connections is done
-          parameters: sanitizedParameterData,
           kind: workflowKind,
         };
 
@@ -138,11 +139,11 @@ export const TemplatesStandaloneDesigner = () => {
             appId,
             workflowNameToUse,
             workflow,
-            undefined,
+            updatedConnectionsData,
             updatedParametersData,
+            updatedSettingProperties,
             undefined,
-            undefined,
-            callBack,
+            () => {},
             true
           );
         } catch (error) {
@@ -154,8 +155,22 @@ export const TemplatesStandaloneDesigner = () => {
     }
   };
 
+  const addConnectionDataInternal = async (connectionAndSetting: ConnectionAndAppSetting): Promise<void> => {
+    addConnectionInJson(connectionAndSetting, connectionsData ?? {});
+    addOrUpdateAppSettings(connectionAndSetting.settings, settingsData?.properties ?? {});
+  };
+
   const services = useMemo(
-    () => getServices(connectionsData ?? {}, workflowAppData as WorkflowApp, tenantId, objectId, canonicalLocation),
+    () =>
+      getServices(
+        isConsumption,
+        connectionsData ?? {},
+        workflowAppData as WorkflowApp,
+        addConnectionDataInternal,
+        tenantId,
+        objectId,
+        canonicalLocation
+      ),
     // eslint-disable-next-line react-hooks/exhaustive-deps
     [connectionsData, settingsData, workflowAppData, tenantId, canonicalLocation]
   );
@@ -176,7 +191,75 @@ export const TemplatesStandaloneDesigner = () => {
             isConsumption={isConsumption}
             existingWorkflowName={existingWorkflowName}
           >
-            <TemplatesDesigner createWorkflowCall={createWorkflowCall} />
+            <div
+              style={{
+                margin: '20px',
+              }}
+            >
+              <TemplateFilters
+                connectors={[
+                  {
+                    value: 'azureaisearch',
+                    displayName: 'Azure AI Search',
+                  },
+                  {
+                    value: 'openai',
+                    displayName: 'Open AI',
+                  },
+                  {
+                    value: 'sql',
+                    displayName: 'SQL',
+                  },
+                  {
+                    value: 'amazon',
+                    displayName: 'Amazon',
+                  },
+                ]}
+                detailFilters={{
+                  By: {
+                    displayName: 'By',
+                    items: [
+                      {
+                        value: 'Microsoft',
+                        displayName: 'Microsoft',
+                      },
+                      {
+                        value: 'Other',
+                        displayName: 'Other',
+                      },
+                    ],
+                  },
+                  Type: {
+                    displayName: 'Type',
+                    items: [
+                      {
+                        value: 'Workflow',
+                        displayName: 'Workflow',
+                      },
+                      {
+                        value: 'Other',
+                        displayName: 'Other',
+                      },
+                    ],
+                  },
+                  Industry: {
+                    displayName: 'Categories',
+                    items: [
+                      {
+                        value: 'Analytics',
+                        displayName: 'Analytics',
+                      },
+                      {
+                        value: 'IT',
+                        displayName: 'IT',
+                      },
+                    ],
+                  },
+                }}
+              />
+              <br />
+              <TemplatesDesigner createWorkflowCall={createWorkflowCall} />
+            </div>
           </TemplatesDataProvider>
         </TemplatesDesignerProvider>
       ) : null}
@@ -188,8 +271,10 @@ const apiVersion = '2020-06-01';
 const httpClient = new HttpClient();
 
 const getServices = (
+  isConsumption: boolean,
   connectionsData: ConnectionsData,
   workflowApp: WorkflowApp | undefined,
+  addConnection: (data: ConnectionAndAppSetting) => Promise<void>,
   tenantId: string | undefined,
   objectId: string | undefined,
   location: string
@@ -215,6 +300,7 @@ const getServices = (
     },
     workflowAppDetails: { appName, identity: workflowApp?.identity as any },
     readConnections: () => Promise.resolve(connectionsData),
+    writeConnection: addConnection as any,
   });
   const gatewayService = new BaseGatewayService({
     baseUrl: armUrl,
@@ -238,16 +324,139 @@ const getServices = (
     tenantId,
     objectId,
   });
+  const operationManifestService = new StandardOperationManifestService(defaultServiceParams);
   const workflowService: IWorkflowService = {
     getCallbackUrl: () => Promise.resolve({} as any),
     getAppIdentity: () => workflowApp?.identity as any,
   };
+
+  const templateService = isConsumption
+    ? undefined
+    : new StandardTemplateService({
+        baseUrl: armUrl,
+        appId: siteResourceId,
+        httpClient,
+        apiVersions: {
+          subscription: apiVersion,
+          gateway: '2018-11-01',
+        },
+        openBladeAfterCreate: () => {
+          console.log('Open blade after create');
+        },
+      });
 
   return {
     connectionService,
     gatewayService,
     tenantService,
     oAuthService,
+    operationManifestService,
+    templateService,
     workflowService,
   };
+};
+
+const replaceWithWorkflowName = (content: string, workflowName: string) => content.replaceAll(workflowIdentifier, workflowName);
+
+const updateConnectionsDataWithNewConnections = async (
+  originalConnectionsData: ConnectionsData,
+  settingProperties: Record<string, string>,
+  connections: ConnectionMapping,
+  workflowJsonString: string,
+  workflowName: string
+): Promise<{ connectionsData: ConnectionsData; settingProperties: Record<string, string>; workflowJsonString: string }> => {
+  const { references, mapping } = connections;
+  let updatedSettings = { ...settingProperties };
+  const updatedConnectionsData = { ...originalConnectionsData };
+  let updatedWorkflowJsonString = workflowJsonString;
+  let updatedConnectionsJsonString = JSON.stringify(updatedConnectionsData);
+  const referencesToProcess: string[] = [];
+
+  for (const connectionKey of Object.keys(mapping)) {
+    const referenceKey = mapping[connectionKey];
+    if (connectionKey === referenceKey) {
+      referencesToProcess.push(referenceKey);
+    } else {
+      updatedWorkflowJsonString = updatedWorkflowJsonString.replaceAll(connectionKey, referenceKey);
+    }
+  }
+
+  if (referencesToProcess.length) {
+    const newManagedApiConnections = updatedConnectionsData.managedApiConnections ?? {};
+    const referencesToNormalize: string[] = [];
+    await Promise.all(
+      referencesToProcess.map(async (referenceKey) => {
+        const reference = references[referenceKey];
+        const normalizedReferenceKey = replaceWithWorkflowName(referenceKey, workflowName);
+        if (isArmResourceId(reference?.connection?.id)) {
+          // Managed API Connection
+          const {
+            api: { id: apiId },
+            connection: { id: connectionId },
+            connectionProperties,
+          } = reference;
+          const connection = await getConnectionStandard(connectionId);
+          const userIdentity = connectionProperties?.authentication?.identity;
+          const newConnectionObj = {
+            api: { id: apiId },
+            connection: { id: connectionId },
+            authentication: {
+              type: 'ManagedServiceIdentity',
+              ...optional('identity', userIdentity),
+            },
+            connectionRuntimeUrl: connection?.properties?.connectionRuntimeUrl ?? '',
+            connectionProperties,
+          };
+
+          newManagedApiConnections[normalizedReferenceKey] = newConnectionObj;
+        } else {
+          referencesToNormalize.push(referenceKey);
+        }
+
+        updatedWorkflowJsonString = updatedWorkflowJsonString.replaceAll(referenceKey, normalizedReferenceKey);
+      })
+    );
+    updatedConnectionsData.managedApiConnections = newManagedApiConnections;
+
+    updatedConnectionsJsonString = JSON.stringify(updatedConnectionsData);
+    for (const referenceKey of referencesToNormalize) {
+      const normalizedReferenceKey = replaceWithWorkflowName(referenceKey, workflowName);
+      const escapedReferenceKeyForSettings = escapeSpecialChars(referenceKey);
+      const normalizedEscapedReferenceKeyForSettings = replaceWithWorkflowName(escapedReferenceKeyForSettings, workflowName);
+
+      updatedConnectionsJsonString = updatedConnectionsJsonString.replaceAll(referenceKey, normalizedReferenceKey);
+      updatedConnectionsJsonString = updatedConnectionsJsonString.replaceAll(
+        escapedReferenceKeyForSettings,
+        normalizedEscapedReferenceKeyForSettings
+      );
+      updatedSettings = updateSettingsKeyWithWorkflowName(
+        updatedSettings,
+        escapedReferenceKeyForSettings,
+        normalizedEscapedReferenceKeyForSettings
+      );
+    }
+  }
+
+  return {
+    connectionsData: JSON.parse(updatedConnectionsJsonString),
+    settingProperties: updatedSettings,
+    workflowJsonString: updatedWorkflowJsonString,
+  };
+};
+
+const updateSettingsKeyWithWorkflowName = (
+  settingProperties: Record<string, string>,
+  referenceKey: string,
+  normalizedReferenceKey: string
+): Record<string, string> => {
+  const updatedSettings = { ...settingProperties };
+  for (const key of Object.keys(settingProperties)) {
+    if (key.includes(referenceKey)) {
+      const normalizedKey = key.replace(referenceKey, normalizedReferenceKey);
+      updatedSettings[normalizedKey] = settingProperties[key];
+      delete updatedSettings[key];
+    }
+  }
+
+  return updatedSettings;
 };