+++ libs/designer/src/lib/core/templates/TemplatesDesignerContext.tsx
@@ -8,7 +8,9 @@ import type {
   ITenantService,
   ILoggerService,
   IOAuthService,
+  ITemplateService,
   IWorkflowService,
+  IOperationManifestService,
 } from '@microsoft/logic-apps-shared';
 import { createContext } from 'react';
 
@@ -18,6 +20,7 @@ export interface TemplatesDesignerContext {
 
 export interface TemplateServiceOptions {
   connectionService: IConnectionService;
+  operationManifestService: IOperationManifestService;
   workflowService: IWorkflowService;
   gatewayService?: IGatewayService;
   tenantService?: ITenantService;
@@ -27,6 +30,7 @@ export interface TemplateServiceOptions {
   functionService?: IFunctionService;
   appServiceService?: IAppServiceService;
   connectionParameterEditorService?: IConnectionParameterEditorService;
+  templateService?: ITemplateService;
 }
 
 export const TemplatesWrappedContext = createContext<TemplatesDesignerContext | null>(null);