+++ libs/designer/src/lib/core/utils/parameters/dynamicdata.ts
@@ -280,6 +280,7 @@ export async function getDynamicInputsFromSchema(
     useAliasedIndexing: true,
     excludeAdvanced: false,
     excludeInternal: false,
+    includeParentObject: true,
   };
   const schemaProperties = new SchemaProcessor(processorOptions).getSchemaProperties(schema);
   let dynamicInputs: InputParameter[] = schemaProperties.map((schemaProperty) => ({
@@ -313,7 +314,7 @@ export async function getDynamicInputsFromSchema(
 
   if (!operationDefinition) {
     loadParameterValuesFromDefault(map(dynamicInputs, 'key'));
-    return dynamicInputs;
+    return removeParentObjectInputsIfNotNeeded(dynamicInputs);
   }
 
   if (!schemaProperties.length) {
@@ -542,7 +543,7 @@ async function getManagedIdentityRequestProperties(
   return managedIdentityRequestProperties;
 }
 
-function getManifestBasedInputParameters(
+export function getManifestBasedInputParameters(
   dynamicInputs: InputParameter[],
   dynamicParameter: InputParameter,
   allInputKeys: string[],
@@ -601,6 +602,8 @@ function getManifestBasedInputParameters(
     }
   }
 
+  result = removeParentObjectInputsIfNotNeeded(result);
+
   if (
     !operationDefinition.metadata?.noUnknownParametersWithManifest &&
     stepInputs !== undefined &&
@@ -623,6 +626,26 @@ function getManifestBasedInputParameters(
   }
 
   return result;
+}
+
+function removeParentObjectInputsIfNotNeeded(inputs: InputParameter[]): InputParameter[] {
+  const objectInputKeysWithExpressionValues = inputs
+    .filter((input) => input.type === Constants.SWAGGER.TYPE.OBJECT && isTemplateExpression(input.value))
+    .map((input) => input.key);
+
+  const filteredInputs = inputs.filter((input) => {
+    if (input.type !== Constants.SWAGGER.TYPE.OBJECT) {
+      return !objectInputKeysWithExpressionValues.some((parentInputKey) => input.key.startsWith(`${parentInputKey}.`));
+    }
+
+    return isTemplateExpression(input.value) || !objectHasLeafProperties(inputs, input.key);
+  });
+
+  return filteredInputs;
+}
+
+function objectHasLeafProperties(allInputs: InputParameter[], key: string): boolean {
+  return allInputs.some((input) => input.key.startsWith(`${key}.`));
 }
 
 function loadUnknownManifestBasedParameters(
@@ -705,12 +728,14 @@ function getSwaggerBasedInputParameters(
     propertyNames,
     getObjectPropertyValue(operationDefinition.inputs, propertyNames)
   );
-  const dynamicInputParameters = loadInputValuesFromDefinition(
+  let dynamicInputParameters = loadInputValuesFromDefinition(
     dynamicInputDefinition as Record<string, any>,
     isNested ? [dynamicParameter] : inputs,
     operationPath,
     basePath as string
   );
+
+  dynamicInputParameters = removeParentObjectInputsIfNotNeeded(dynamicInputParameters);
 
   if (isNested) {
     const parameter = first((inputParameter) => inputParameter.key === key, dynamicInputParameters);