+++ libs/designer/src/lib/ui/templates/connections/connector.tsx
@@ -1,13 +1,17 @@
 import type { IImageStyles, IImageStyleProps, IStyleFunctionOrObject } from '@fluentui/react';
-import { Icon, ImageFit, Shimmer, ShimmerElementType, Spinner, SpinnerSize, Text, css } from '@fluentui/react';
+import { Icon, Shimmer, ShimmerElementType, Spinner, SpinnerSize, Text, css } from '@fluentui/react';
 import { useConnectorOnly } from '../../../core/state/connection/connectionSelector';
-import type { Connector, Template } from '@microsoft/logic-apps-shared';
+import type { Template } from '@microsoft/logic-apps-shared';
+import type { IntlShape } from 'react-intl';
 import { useIntl } from 'react-intl';
 import { getConnectorAllCategories } from '@microsoft/designer-ui';
 import { useConnectionsForConnector } from '../../../core/queries/connections';
 import { getConnectorResources } from '../../../core/templates/utils/helper';
+import { useEffect } from 'react';
+import type { ConnectorInfo } from '../../../core/templates/utils/queries';
+import { useConnectorInfo } from '../../../core/templates/utils/queries';
 
-const iconStyles = {
+export const iconStyles = {
   root: {
     width: 20,
     height: 20,
@@ -16,40 +20,46 @@ const iconStyles = {
 
 export const ConnectorIcon = ({
   connectorId,
-  styles,
+  operationId,
   classes,
-}: { connectorId: string; classes: Record<string, string>; styles?: IStyleFunctionOrObject<IImageStyleProps, IImageStyles> }) => {
-  const { data: connector, isLoading, isError } = useConnectorOnly(connectorId);
+}: {
+  connectorId: string;
+  classes: Record<string, string>;
+  operationId?: string;
+  styles?: IStyleFunctionOrObject<IImageStyleProps, IImageStyles>;
+}) => {
+  const { data: connector, isLoading, isError } = useConnectorInfo(connectorId, operationId);
   if (!connector) {
     return isLoading ? <Spinner size={SpinnerSize.small} /> : isError ? <Icon iconName="Error" /> : <Icon iconName="Unknown" />;
   }
 
   return (
     <div className={classes['root']}>
-      <Icon
-        className={classes['icon']}
-        imageProps={{
-          styles: styles ?? iconStyles,
-          src: connector.properties.iconUrl ?? connector.properties.iconUri,
-          imageFit: ImageFit.centerContain,
-        }}
-      />
+      <img className={classes['icon']} src={connector?.iconUrl} />
     </div>
   );
 };
 
 export const ConnectorIconWithName = ({
   connectorId,
+  operationId,
   classes,
   showProgress,
   onConnectorLoaded,
 }: {
   connectorId: string;
   classes: Record<string, string>;
+  operationId?: string;
   showProgress?: boolean;
-  onConnectorLoaded?: (connector: Connector) => void;
+  onConnectorLoaded?: (connector: ConnectorInfo) => void;
 }) => {
-  const { data: connector, isLoading } = useConnectorOnly(connectorId);
+  const { data: connector, isLoading } = useConnectorInfo(connectorId, operationId);
+
+  useEffect(() => {
+    if (onConnectorLoaded && connector) {
+      onConnectorLoaded(connector);
+    }
+  }, [connector, onConnectorLoaded]);
 
   if (showProgress && isLoading) {
     return (
@@ -70,14 +80,10 @@ export const ConnectorIconWithName = ({
     );
   }
 
-  if (onConnectorLoaded && connector) {
-    onConnectorLoaded(connector);
-  }
-
   return (
     <div className={classes['root']}>
-      <img className={classes['icon']} src={connector?.properties.iconUrl ?? connector?.properties.iconUri} />
-      <Text className={classes['text']}>{connector?.properties.displayName}</Text>
+      <img className={classes['icon']} src={connector?.iconUrl} />
+      <Text className={classes['text']}>{connector?.displayName}</Text>
     </div>
   );
 };
@@ -103,13 +109,30 @@ export const ConnectorWithDetails = ({ connectorId, kind }: Template.Connection)
   const text = getConnectorResources(intl);
   return (
     <div className="msla-template-connector">
-      <ConnectorIcon
-        connectorId={connectorId}
-        classes={{ root: 'msla-template-connector-box', icon: 'msla-template-connector-icon' }}
-        styles={{ root: { width: 50, height: 50 }, image: { width: 'calc(60%)', height: 'calc(60%)' } }}
-      />
+      {isLoading ? (
+        <Shimmer
+          className="msla-template-connector-box"
+          shimmerElements={[{ type: ShimmerElementType.line, height: 50, verticalAlign: 'bottom', width: '100%' }]}
+          size={SpinnerSize.xSmall}
+        />
+      ) : (
+        <ConnectorIcon
+          connectorId={connectorId}
+          classes={{ root: 'msla-template-connector-box', icon: 'msla-template-connector-icon' }}
+          styles={{ root: { width: 50, height: 50 } }}
+        />
+      )}
       <div className="msla-template-connector-details">
-        <div className="msla-template-connector-name">{connector.properties.displayName}</div>
+        {isLoading ? (
+          <Shimmer
+            className="msla-template-connector-name"
+            style={{ width: '70%', marginTop: 10 }}
+            shimmerElements={[{ type: ShimmerElementType.line, height: 12, verticalAlign: 'bottom', width: '100%' }]}
+            size={SpinnerSize.xSmall}
+          />
+        ) : (
+          <div className="msla-template-connector-name">{connector.properties.displayName}</div>
+        )}
         <div className="msla-template-connector-type">
           <Text style={textStyles.connectorSubDetails} className="msla-template-card-tag">
             {allCategories[kind ?? ''] ?? kind}
@@ -131,6 +154,35 @@ export const ConnectorWithDetails = ({ connectorId, kind }: Template.Connection)
           )}
         </div>
       </div>
+    </div>
+  );
+};
+
+export const ConnectorConnectionStatus = ({
+  connectorId,
+  hasConnection,
+  intl,
+}: { connectorId: string; hasConnection: boolean; intl: IntlShape }) => {
+  const { data: connector, isLoading } = useConnectorOnly(connectorId);
+  const texts = getConnectorResources(intl);
+  const fontStyle = { color: hasConnection ? '#50821b' : '#8b8b8b' };
+
+  return (
+    <div className="msla-templates-tab-review-section-details">
+      {isLoading ? (
+        <div className="msla-templates-tab-review-section-details-title">
+          <Shimmer
+            style={{ width: '70%', marginTop: 5 }}
+            shimmerElements={[{ type: ShimmerElementType.line, height: 10, verticalAlign: 'bottom', width: '100%' }]}
+            size={SpinnerSize.xSmall}
+          />
+        </div>
+      ) : (
+        <Text className="msla-templates-tab-review-section-details-title">{connector?.properties.displayName}</Text>
+      )}
+      <Text style={fontStyle} className="msla-templates-tab-review-section-details-value">
+        {hasConnection ? texts.connected : texts.notConnected}
+      </Text>
     </div>
   );
 };