+++ libs/designer/src/lib/ui/templates/index.tsx
@@ -2,4 +2,5 @@ import './styles.less';
 import '../styles.less';
 
 export * from './TemplatesDesigner';
+export * from './filters/templateFilters';
 export * from './connections/displayConnections';