+++ libs/logic-apps-shared/src/designer-client-services/lib/standard/operationmanifest.ts
@@ -2,6 +2,7 @@ import type { OperationInfo, OperationManifest } from '../../../utils/src';
 import { ConnectionType, equals } from '../../../utils/src';
 import { BaseOperationManifestService } from '../base';
 import { getBuiltInOperationInfo, isBuiltInOperation, supportedBaseManifestObjects } from '../base/operationmanifest';
+import { getHybridAppBaseRelativeUrl, isHybridLogicApp } from './hybrid';
 
 export class StandardOperationManifestService extends BaseOperationManifestService {
   override async getOperationInfo(definition: any, isTrigger: boolean): Promise<OperationInfo> {
@@ -38,10 +39,21 @@ export class StandardOperationManifestService extends BaseOperationManifestServi
     };
 
     try {
-      const response = await httpClient.get<any>({
-        uri: `${baseUrl}/operationGroups/${connectorName}/operations/${operationName}`,
-        queryParameters,
-      });
+      let response = null;
+      if (isHybridLogicApp(baseUrl)) {
+        response = await httpClient.post<any, null>({
+          uri: `${getHybridAppBaseRelativeUrl(baseUrl.split('hostruntime')[0])}/invoke?api-version=2024-02-02-preview`,
+          headers: {
+            'x-ms-logicapps-proxy-path': `/runtime/webhooks/workflow/api/management/operationGroups/${connectorName}/operations/${operationName}?$expand=properties/manifest`,
+            'x-ms-logicapps-proxy-method': 'GET',
+          },
+        });
+      } else {
+        response = await httpClient.get<any>({
+          uri: `${baseUrl}/operationGroups/${connectorName}/operations/${operationName}`,
+          queryParameters,
+        });
+      }
 
       const {
         properties: { brandColor, description, iconUri, manifest, operationType, api },